requirements.txt
torch>=2.0.0
torch-geometric>=2.3.0
rdkit>=2023.3.1
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0
pyyaml>=6.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
umap-learn>=0.5.3
pytest>=7.4.0